+++ src/main/java/it/gov/pagopa/apiconfig/cache/controller/FdrCacheController.java
@@ -29,127 +29,9 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/stakeholders/fdr/cache/schemas")
 @Validated
 @Slf4j
-public class FdrCacheController {
-
-  private String stakeholder = "fdr";
-
-  private ConfigDataV1 cfgDataV1 = null;
-
-  @Autowired private ConfigService configService;
-
-  @PostConstruct
-  private void loadCacheFromRedis() {
-    cfgDataV1 = configService.loadFromRedis(stakeholder);
-  }
-
-  @Operation(
-      summary = "Get selected key of fdr v1 config",
-      security = {@SecurityRequirement(name = "ApiKey")},
-      tags = {
-        "FdrCache",
-      })
-  @ApiResponses(
-      value = {
-        @ApiResponse(
-            responseCode = "200",
-            description = "OK",
-            content =
-                @Content(
-                    mediaType = MediaType.APPLICATION_JSON_VALUE,
-                    schema = @Schema(implementation = ConfigDataV1.class))),
-        @ApiResponse(
-            responseCode = "400",
-            description = "Bad Request",
-            content =
-                @Content(
-                    mediaType = MediaType.APPLICATION_JSON_VALUE,
-                    schema = @Schema(implementation = ProblemJson.class))),
-        @ApiResponse(
-            responseCode = "401",
-            description = "Unauthorized",
-            content = @Content(schema = @Schema())),
-        @ApiResponse(
-            responseCode = "403",
-            description = "Forbidden",
-            content = @Content(schema = @Schema())),
-        @ApiResponse(
-            responseCode = "429",
-            description = "Too many requests",
-            content = @Content(schema = @Schema())),
-        @ApiResponse(
-            responseCode = "500",
-            description = "Service unavailable",
-            content =
-                @Content(
-                    mediaType = MediaType.APPLICATION_JSON_VALUE,
-                    schema = @Schema(implementation = ProblemJson.class)))
-      })
-  @GetMapping(
-      value = "/v1",
-      produces = {MediaType.APPLICATION_JSON_VALUE})
-  public ResponseEntity<ConfigDataV1> cache(
-      @RequestParam Optional<Boolean> refresh, @RequestParam Optional<NodeCacheKey[]> keys)
-      throws IOException {
-    boolean cacheV1InProgress = configService.getCacheV1InProgress(stakeholder);
-    if (refresh.orElse(false) || cfgDataV1 == null) {
-      if (cacheV1InProgress) {
-        return ResponseEntity.status(HttpStatus.SERVICE_UNAVAILABLE).build();
-      }
-      cfgDataV1 = configService.newCacheV1(stakeholder, keys);
-    }
-    return ResponseEntity.ok(cfgDataV1);
-  }
-
-  @Operation(
-      summary = "Get last fdr v1 cache version",
-      security = {@SecurityRequirement(name = "ApiKey")},
-      tags = {
-        "FdrCache",
-      })
-  @ApiResponses(
-      value = {
-        @ApiResponse(
-            responseCode = "200",
-            description = "OK",
-            content =
-                @Content(
-                    mediaType = MediaType.APPLICATION_JSON_VALUE,
-                    schema = @Schema(implementation = CacheVersion.class))),
-        @ApiResponse(
-            responseCode = "400",
-            description = "Bad Request",
-            content =
-                @Content(
-                    mediaType = MediaType.APPLICATION_JSON_VALUE,
-                    schema = @Schema(implementation = ProblemJson.class))),
-        @ApiResponse(
-            responseCode = "401",
-            description = "Unauthorized",
-            content = @Content(schema = @Schema())),
-        @ApiResponse(
-            responseCode = "403",
-            description = "Forbidden",
-            content = @Content(schema = @Schema())),
-        @ApiResponse(
-            responseCode = "404",
-            description = "Not Found",
-            content = @Content(schema = @Schema())),
-        @ApiResponse(
-            responseCode = "429",
-            description = "Too many requests",
-            content = @Content(schema = @Schema())),
-        @ApiResponse(
-            responseCode = "500",
-            description = "Service unavailable",
-            content =
-                @Content(
-                    mediaType = MediaType.APPLICATION_JSON_VALUE,
-                    schema = @Schema(implementation = ProblemJson.class)))
-      })
-  @GetMapping(
-      value = "/v1/id",
-      produces = {MediaType.APPLICATION_JSON_VALUE})
-  public ResponseEntity<CacheVersion> idV1() {
-    return ResponseEntity.ok(configService.getCacheV1Id(stakeholder));
+public class FdrCacheController extends CacheController{
+  @Override
+  String stakeholder() {
+    return "fdr";
   }
 }
